requirements.txt
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
cycler==0.10.0
django-crispy-forms==1.7.2
django-elasticsearch-dsl==6.4.2
django-javascript-settings==1.1.2
django-popup-view-field==0.5.0
elasticsearch==6.4.0
elasticsearch-dsl==6.4.0
ipaddress==1.0.22
kiwisolver==1.1.0
kombu==4.6.3
matplotlib==2.2.4
numpy==1.16.2
Pillow==6.0.0
psycopg2-binary==2.8.2
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
urllib3==1.25.3
vine==1.3.0